Bluetooth: Rename list.h to avoid name conflicts.
bug: v/46057

Rootcause:The header files are accessible by all modules all modules, which makes it easy for the list.h to conflict with other file names, so the file name should be renamed.
