+++ service/profiles/spp/spp_service.c
@@ -24,11 +24,11 @@
 #include "bt_addr.h"
 #include "bt_debug.h"
 #include "bt_device.h"
+#include "bt_list.h"
 #include "bt_profile.h"
 #include "bt_uuid.h"
 #include "euv_pty.h"
 #include "index_allocator.h"
-#include "list.h"
 #include "openpty.h"
 #include "power_manager.h"
 #include "sal_spp_interface.h"
